# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'My first API'
  contact:
    email: support@example.com
  version: 1.0.0
paths:
...
  /users/me:
    get:
      tags:
        - users
      summary: 'Retorna o usuário logado'
      operationId: GetAuthUser
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
    put:
      tags:
        - users
      summary: 'Atualiza o usuário'
      operationId: UpdateAuthUser
      parameters:
        -
          name: first_name
          in: query
          description: 'Primeiro nome'
          required: false
          schema:
            type: string
        -
          name: last_name
          in: query
          description: 'Ultimo nome'
          required: false
          schema:
            type: string
        -
          name: document_number
          in: query
          description: 'Documento CPF'
          required: false
          schema:
            type: string
        -
          name: address_street
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_number
          in: query
          description: 'Numero da casa do rapaz'
          required: false
          schema:
            type: string
        -
          name: address_city
          in: query
          description: 'Cidade onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_state
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do cidadão'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  /users:
    get:
      tags:
        - users
      summary: 'Lista todas os usuários'
      operationId: GetUsers
      parameters:
        -
          name: includes
          in: query
          description: 'Faz o include das relações '
          required: false
          schema:
            type: array
            items:
              type: string
        -
          name: first_name
          in: query
          description: 'Nome do usuário'
          required: false
          schema:
            type: string
        -
          name: last_name
          in: query
          description: 'Sobrenome do usuário'
          required: false
          schema:
            type: string
        -
          name: document_number
          in: query
          description: CPF/CNPJ
          required: false
          schema:
            type: string
        -
          name: is_admin
          in: query
          description: 'Status de Administrador'
          required: false
          schema:
            type: number
            enum:
              - 0
              - 1
      responses:
        '200':
          description: ...
    put:
      tags:
        - users
      summary: 'Atualiza um usuário'
      operationId: UpdateUser
      parameters:
        -
          name: document_number
          in: query
          description: 'Documento CPF'
          required: false
          schema:
            type: string
        -
          name: first_name
          in: query
          description: 'Primeiro nome'
          required: false
          schema:
            type: string
        -
          name: last_name
          in: query
          description: 'Ultimo nome'
          required: false
          schema:
            type: string
        -
          name: address_street
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_number
          in: query
          description: 'Numero da casa do rapaz'
          required: false
          schema:
            type: string
        -
          name: address_city
          in: query
          description: 'Cidade onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_state
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do cidadão'
          required: false
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do cidadão'
          required: false
          schema:
            type: string
        -
          name: is_admin
          in: query
          description: 'Status de Administrador'
          required: false
          schema:
            type: number
            enum:
              - 0
              - 1
        -
          name: facebook_id
          in: query
          description: 'ID do Facebook'
          required: false
          schema:
            type: number
        -
          name: google_id
          in: query
          description: 'ID do Google'
          required: false
          schema:
            type: number
      responses:
        '200':
          description: ...
    post:
      tags:
        - users
      summary: 'Cria um novo usuário'
      operationId: StoreUser
      parameters:
        -
          name: first_name
          in: query
          description: 'Primeiro nome'
          required: true
          schema:
            type: string
        -
          name: last_name
          in: query
          description: 'Ultimo nome'
          required: true
          schema:
            type: string
        -
          name: document_number
          in: query
          description: 'Documento CPF'
          required: false
          schema:
            type: string
        -
          name: address_street
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_number
          in: query
          description: 'Numero da casa do rapaz'
          required: false
          schema:
            type: string
        -
          name: address_city
          in: query
          description: 'Cidade onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_state
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do cidadão'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do cidadão'
          required: true
          schema:
            type: string
        -
          name: is_admin
          in: query
          description: 'Status de Administrador'
          required: true
          schema:
            type: number
            enum:
              - 0
              - 1
      responses:
        '200':
          description: ...
  '/users/{id}':
    get:
      tags:
        - users
      summary: 'Lista um usuário'
      operationId: GetUser
      parameters:
        -
          name: id
          in: path
          description: 'Id do usuário'
          required: true
          schema:
            type: number
      responses:
        '200':
          description: ...
    delete:
      tags:
        - users
      summary: 'Apaga um usuário'
      operationId: DeleteUser
      parameters:
        -
          name: id
          in: path
          description: 'Id do usuário'
          required: true
          schema:
            type: number
      responses:
        '200':
          description: ...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      description: 'Token do client para consumo da API'
      name: Authorization
      in: header
      scheme: bearer